+++ src/app/modules/user/home.page/components/skeleton/board-list.skeleton.tsx
@@ -1,0 +1,9 @@
+import { Skeleton } from "@/app/atomic";
+
+export function BoardListSkeleton() {
+  return (
+    <>
+      <Skeleton />
+    </>
+  );
+}
